fix: removed duplicated code
