+++ deployment/united-manufacturing-hub/values.yaml
@@ -641,7 +641,7 @@ kafkatopostgresql:
     repository: unitedmanufacturinghub/kafka-to-postgresql
     pullPolicy: IfNotPresent
     #tag: latest
-  replicas: 1
+  replicas: 3
   storageRequest: 1Gi
   listenTopic:
     # All topics except processValue
@@ -2053,6 +2053,7 @@ kafka:
   logRetentionHours: 168
   logSegmentBytes: _10485760 # 10MiB
   logRetentionBytes: _26214400 # 25MiB
+  numPartitions: 6
 
 ##### CONFIG FOR REDIS #####
 redis:
