+++ golang/cmd/mqtt-kafka-bridge/main.go
@@ -10,8 +10,6 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
-	"regexp"
-	"strings"
 	"syscall"
 	"time"
 )
@@ -126,23 +124,4 @@ func ShutdownApplicationGraceful() {
 	// Gracefully exit.
 	// (Use runtime.GoExit() if you need to call defers)
 	os.Exit(0)
-}
-
-var validKafkaTopicRegex, _ = regexp.Compile(`^[a-zA-Z\d\._\-]+$`)
-
-func MqttTopicToKafka(MqttTopicName string) (KafkaTopicName string) {
-
-	MqttTopicName = strings.TrimSpace(MqttTopicName)
-	MqttTopicName = strings.ReplaceAll(MqttTopicName, "/", ".")
-	if !validKafkaTopicRegex.Match([]byte(MqttTopicName)) {
-		zap.S().Errorf("Invalid MQTT->Kafka topic name: %s", MqttTopicName)
-	}
-	return MqttTopicName
-}
-func KafkaTopicToMqtt(KafkaTopicName string) (MqttTopicName string) {
-	if strings.Contains(KafkaTopicName, "/") {
-		zap.S().Errorf("Illegal MQTT->Kafka Topic name: %s", KafkaTopicName)
-	}
-	KafkaTopicName = strings.TrimSpace(KafkaTopicName)
-	return strings.ReplaceAll(KafkaTopicName, ".", "/")
 }
